Renamed the static_classes folder to utitilies class. And utilities folder renamed to mocks
