+++ src/test/java/apimatic/core/RequestBuilderTest.java
@@ -25,8 +25,8 @@ import org.mockito.Captor;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnit;
 import org.mockito.junit.MockitoRule;
+import apimatic.core.mocks.MockCoreConfig;
 import apimatic.core.models.Employee;
-import apimatic.core.utilities.MockCoreConfig;
 import io.apimatic.core.ApiCall;
 import io.apimatic.core.HttpRequest;
 import io.apimatic.core.authentication.HeaderAuth;
