+++ src/test/java/apimatic/core/ResponseHandlerTest.java
@@ -19,8 +19,8 @@ import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnit;
 import org.mockito.junit.MockitoRule;
 import apimatic.core.exceptions.GlobalTestException;
+import apimatic.core.mocks.MockCoreConfig;
 import apimatic.core.models.TestModel;
-import apimatic.core.utilities.MockCoreConfig;
 import io.apimatic.core.ApiCall;
 import io.apimatic.core.ErrorCase;
 import io.apimatic.core.ResponseHandler;
