+++ src/test/java/apimatic/core/utilities/LocalDateTimeHelperTest.java
@@ -1,4 +1,4 @@
-package apimatic.core.static_classes;
+package apimatic.core.utilities;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNull;
@@ -15,7 +15,7 @@ import com.fasterxml.jackson.databind.JsonDeserializer;
 import com.fasterxml.jackson.databind.JsonSerializer;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.module.SimpleModule;
-import apimatic.core.utilities.TestDateTimeHelper;
+import apimatic.core.mocks.TestDateTimeHelper;
 import io.apimatic.core.utilities.LocalDateTimeHelper;
 
 public class LocalDateTimeHelperTest {
